+++ source/Private/Assert-SetupActionProperties.ps1
@@ -26,6 +26,7 @@
 #>
 function Assert-SetupActionProperties
 {
+    [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseSingularNouns', '', Justification='The command uses plural noun to describe that it contain a collection of asserts.')]
     [CmdletBinding()]
     param
     (