+++ source/Public/Complete-SqlDscFailoverCluster.ps1
@@ -15,6 +15,117 @@
         Specifies the path where to find the SQL Server installation media. On this
         path the SQL Server setup executable must be found.
 
+    .PARAMETER Timeout
+        Specifies how long to wait for the setup process to finish. Default value
+        is `7200` seconds (2 hours). If the setup process does not finish before
+        this time, an exception will be thrown.
+
+    .PARAMETER Force
+        If specified the command will not ask for confirmation. Same as if Confirm:$false
+        is used.
+
+    .PARAMETER InstanceName
+        See the notes section for more information.
+
+    .PARAMETER Enu
+        See the notes section for more information.
+
+    .PARAMETER ProductKey
+        See the notes section for more information.
+
+    .PARAMETER ASBackupDir
+        See the notes section for more information.
+
+    .PARAMETER ASCollation
+        See the notes section for more information.
+
+    .PARAMETER ASConfigDir
+        See the notes section for more information.
+
+    .PARAMETER ASDataDir
+        See the notes section for more information.
+
+    .PARAMETER ASLogDir
+        See the notes section for more information.
+
+    .PARAMETER ASTempDir
+        See the notes section for more information.
+
+    .PARAMETER ASServerMode
+        See the notes section for more information.
+
+    .PARAMETER ASSysAdminAccounts
+        See the notes section for more information.
+
+    .PARAMETER ASProviderMSOLAP
+        See the notes section for more information.
+
+    .PARAMETER InstallSqlDataDir
+        See the notes section for more information.
+
+    .PARAMETER SqlBackupDir
+        See the notes section for more information.
+
+    .PARAMETER SecurityMode
+        See the notes section for more information.
+
+    .PARAMETER SAPwd
+        See the notes section for more information.
+
+    .PARAMETER SqlCollation
+        See the notes section for more information.
+
+    .PARAMETER SqlSysAdminAccounts
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbDir
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbLogDir
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbFileCount
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbFileSize
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbFileGrowth
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbLogFileSize
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbLogFileGrowth
+        See the notes section for more information.
+
+    .PARAMETER SqlUserDbDir
+        See the notes section for more information.
+
+    .PARAMETER SqlUserDbLogDir
+        See the notes section for more information.
+
+    .PARAMETER RsInstallMode
+        See the notes section for more information.
+
+    .PARAMETER FailoverClusterGroup
+        See the notes section for more information.
+
+    .PARAMETER FailoverClusterDisks
+        See the notes section for more information.
+
+    .PARAMETER FailoverClusterNetworkName
+        See the notes section for more information.
+
+    .PARAMETER FailoverClusterIPAddresses
+        See the notes section for more information.
+
+    .PARAMETER ConfirmIPDependencyChange
+        See the notes section for more information.
+
+    .PARAMETER ProductCoveredBySA
+        See the notes section for more information.
+
     .LINK
         https://docs.microsoft.com/en-us/sql/database-engine/install-windows/install-sql-server-from-the-command-prompt
 
@@ -28,10 +139,9 @@
         Failover Cluster instance.
 
     .NOTES
-        All parameters has intentionally not been added to this comment-based help
-        since it would take a lot of effort to keep it up to date. Instead there is
-        a link in the comment-based help that points to the SQL Server command line
-        setup documentation which will stay relevant.
+        The parameters are intentionally not described since it would take a lot
+        of effort to keep them up to date. Instead there is a link that points to
+        the SQL Server command line setup documentation which will stay relevant.
 #>
 function Complete-SqlDscFailoverCluster
 {