+++ source/Public/Complete-SqlDscImage.ps1
@@ -19,6 +19,141 @@
         Specifies the path where to find the SQL Server installation media. On this
         path the SQL Server setup executable must be found.
 
+    .PARAMETER Timeout
+        Specifies how long to wait for the setup process to finish. Default value
+        is `7200` seconds (2 hours). If the setup process does not finish before
+        this time, an exception will be thrown.
+
+    .PARAMETER Force
+        If specified the command will not ask for confirmation. Same as if Confirm:$false
+        is used.
+
+    .PARAMETER InstanceName
+        See the notes section for more information.
+
+    .PARAMETER Enu
+        See the notes section for more information.
+
+    .PARAMETER InstanceId
+        See the notes section for more information.
+
+    .PARAMETER PBEngSvcAccount
+        See the notes section for more information.
+
+    .PARAMETER PBEngSvcPassword
+        See the notes section for more information.
+
+    .PARAMETER PBEngSvcStartupType
+        See the notes section for more information.
+
+    .PARAMETER PBStartPortRange
+        See the notes section for more information.
+
+    .PARAMETER PBEndPortRange
+        See the notes section for more information.
+
+    .PARAMETER PBScaleOut
+        See the notes section for more information.
+
+    .PARAMETER ProductKey
+        See the notes section for more information.
+
+    .PARAMETER AgtSvcAccount
+        See the notes section for more information.
+
+    .PARAMETER AgtSvcPassword
+        See the notes section for more information.
+
+    .PARAMETER AgtSvcStartupType
+        See the notes section for more information.
+
+    .PARAMETER BrowserSvcStartupType
+        See the notes section for more information.
+
+    .PARAMETER EnableRanU
+        See the notes section for more information.
+
+    .PARAMETER InstallSqlDataDir
+        See the notes section for more information.
+
+    .PARAMETER SqlBackupDir
+        See the notes section for more information.
+
+    .PARAMETER SecurityMode
+        See the notes section for more information.
+
+    .PARAMETER SAPwd
+        See the notes section for more information.
+
+    .PARAMETER SqlCollation
+        See the notes section for more information.
+
+    .PARAMETER SqlSvcAccount
+        See the notes section for more information.
+
+    .PARAMETER SqlSvcPassword
+        See the notes section for more information.
+
+    .PARAMETER SqlSvcStartupType
+        See the notes section for more information.
+
+    .PARAMETER SqlSysAdminAccounts
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbDir
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbLogDir
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbFileCount
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbFileSize
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbFileGrowth
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbLogFileSize
+        See the notes section for more information.
+
+    .PARAMETER SqlTempDbLogFileGrowth
+        See the notes section for more information.
+
+    .PARAMETER SqlUserDbDir
+        See the notes section for more information.
+
+    .PARAMETER SqlUserDbLogDir
+        See the notes section for more information.
+
+    .PARAMETER FileStreamLevel
+        See the notes section for more information.
+
+    .PARAMETER FileStreamShareName
+        See the notes section for more information.
+
+    .PARAMETER NpEnabled
+        See the notes section for more information.
+
+    .PARAMETER TcpEnabled
+        See the notes section for more information.
+
+    .PARAMETER RsInstallMode
+        See the notes section for more information.
+
+    .PARAMETER RSSvcAccount
+        See the notes section for more information.
+
+    .PARAMETER RSSvcPassword
+        See the notes section for more information.
+
+    .PARAMETER RSSvcStartupType
+        See the notes section for more information.
+
+    .PARAMETER ProductCoveredBySA
+        See the notes section for more information.
+
     .LINK
         https://docs.microsoft.com/en-us/sql/database-engine/install-windows/install-sql-server-from-the-command-prompt
 
@@ -32,10 +167,9 @@
         was prepared using `Install-SqlDscServer` with the parameter `-PrepareImage`.
 
     .NOTES
-        All parameters has intentionally not been added to this comment-based help
-        since it would take a lot of effort to keep it up to date. Instead there is
-        a link in the comment-based help that points to the SQL Server command line
-        setup documentation which will stay relevant.
+        The parameters are intentionally not described since it would take a lot
+        of effort to keep them up to date. Instead there is a link that points to
+        the SQL Server command line setup documentation which will stay relevant.
 #>
 function Complete-SqlDscImage
 {