+++ CHANGELOG.md
@@ -68,6 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add the GitHub App _Pull Request Quantifier_ as an experiment to see if it
     brings any value ([issue #1811](https://github.com/dsccommunity/SqlServerDsc/issues/1811)).
     - Updated thresholds, and label names and colors.
+  - New QA tests for public commands and private functions.
 - SqlDatabase
   - Added compatibility levels for SQL Server 2022 (major version 16).
 - SqlSetup